requirements.txt
redis==2.10.1
sockjs-tornado==1.0.1
tornado==4.0.2
fabric
sphinx
Django==1.8
websocket-client==0.32.0
sphinxcontrib-plantuml
sphinx-autobuild
requests
PySQLPool==0.3.8
django-jsonview==0.4.3
sphinxcontrib-autoanysrc
django-suit==0.2.13
boto
django-tornado
django-celery==3.1.17
django-celery-email==1.0.3
tornado-redis==2.4.18
django-sockjs-tornado
